simplify code quite a bit
