--- notes/notes_7_16_19.txt
@@ -1,1 +1,0 @@
-du/dt = c   train loss=6.93e-7  val_loss=5.4e-6
