--- src/navierstokes/flow_periodic.py
@@ -1,230 +1,0 @@
-"""
-This is a DEMO script. It is intended to be run in isolation. 
-Please see flow.py for a reliable version.
-"""
-# periodic flow requires special computation
-
-import os
-import numpy as np
-
-
-def build_up_b(rho, dt, dx, dy, u, v):
-    b = np.zeros_like(u)
-    b[1:-1, 1:-1] = (rho * (1 / dt * ((u[1:-1, 2:] - u[1:-1, 0:-2]) / (2 * dx) +
-                                      (v[2:, 1:-1] - v[0:-2, 1:-1]) / (2 * dy)) -
-                            ((u[1:-1, 2:] - u[1:-1, 0:-2]) / (2 * dx))**2 -
-                            2 * ((u[2:, 1:-1] - u[0:-2, 1:-1]) / (2 * dy) *
-                                 (v[1:-1, 2:] - v[1:-1, 0:-2]) / (2 * dx))-
-                            ((v[2:, 1:-1] - v[0:-2, 1:-1]) / (2 * dy))**2))
-    
-    # Periodic BC Pressure @ x = 2
-    b[1:-1, -1] = (rho * (1 / dt * ((u[1:-1, 0] - u[1:-1,-2]) / (2 * dx) +
-                                    (v[2:, -1] - v[0:-2, -1]) / (2 * dy)) -
-                          ((u[1:-1, 0] - u[1:-1, -2]) / (2 * dx))**2 -
-                          2 * ((u[2:, -1] - u[0:-2, -1]) / (2 * dy) *
-                               (v[1:-1, 0] - v[1:-1, -2]) / (2 * dx)) -
-                          ((v[2:, -1] - v[0:-2, -1]) / (2 * dy))**2))
-
-    # Periodic BC Pressure @ x = 0
-    b[1:-1, 0] = (rho * (1 / dt * ((u[1:-1, 1] - u[1:-1, -1]) / (2 * dx) +
-                                   (v[2:, 0] - v[0:-2, 0]) / (2 * dy)) -
-                         ((u[1:-1, 1] - u[1:-1, -1]) / (2 * dx))**2 -
-                         2 * ((u[2:, 0] - u[0:-2, 0]) / (2 * dy) *
-                              (v[1:-1, 1] - v[1:-1, -1]) / (2 * dx))-
-                         ((v[2:, 0] - v[0:-2, 0]) / (2 * dy))**2))
-    
-    return b
-
-
-def pressure_poisson(p, dx, dy, b):
-    pn = np.empty_like(p)
-    
-    for q in range(nit):
-        pn = p.copy()
-        p[1:-1, 1:-1] = (((pn[1:-1, 2:] + pn[1:-1, 0:-2]) * dy**2 +
-                          (pn[2:, 1:-1] + pn[0:-2, 1:-1]) * dx**2) /
-                         (2 * (dx**2 + dy**2)) -
-                         dx**2 * dy**2 / (2 * (dx**2 + dy**2)) * b[1:-1, 1:-1])
-
-        # Periodic BC Pressure @ x = 2
-        p[1:-1, -1] = (((pn[1:-1, 0] + pn[1:-1, -2])* dy**2 +
-                        (pn[2:, -1] + pn[0:-2, -1]) * dx**2) /
-                       (2 * (dx**2 + dy**2)) -
-                       dx**2 * dy**2 / (2 * (dx**2 + dy**2)) * b[1:-1, -1])
-
-        # Periodic BC Pressure @ x = 0
-        p[1:-1, 0] = (((pn[1:-1, 1] + pn[1:-1, -1])* dy**2 +
-                       (pn[2:, 0] + pn[0:-2, 0]) * dx**2) /
-                      (2 * (dx**2 + dy**2)) -
-                      dx**2 * dy**2 / (2 * (dx**2 + dy**2)) * b[1:-1, 0])
-        
-        # Wall boundary conditions, pressure
-        p[-1, :] = p[-2, :]  # dp/dy = 0 at y = 2
-        p[0, :] = p[1, :]    # dp/dy = 0 at y = 0
-    
-    return p
-
-
-def navier_stokes_flow(nt, u, v, dt, dx, dy, p, rho, nu, F):
-    # collect propagations for dataset
-    u_list, v_list, p_list = [], [], []
-
-    for n in range(nt):
-        un = u.copy()
-        vn = v.copy()
-
-        b = build_up_b(rho, dt, dx, dy, u, v)
-        p = pressure_poisson_periodic(p, dx, dy, b)
-
-        u[1:-1, 1:-1] = (un[1:-1, 1:-1] -
-                        un[1:-1, 1:-1] * dt / dx * 
-                        (un[1:-1, 1:-1] - un[1:-1, 0:-2]) -
-                        vn[1:-1, 1:-1] * dt / dy * 
-                        (un[1:-1, 1:-1] - un[0:-2, 1:-1]) -
-                        dt / (2 * rho * dx) * 
-                        (p[1:-1, 2:] - p[1:-1, 0:-2]) +
-                        nu * (dt / dx**2 * 
-                        (un[1:-1, 2:] - 2 * un[1:-1, 1:-1] + un[1:-1, 0:-2]) +
-                        dt / dy**2 * 
-                        (un[2:, 1:-1] - 2 * un[1:-1, 1:-1] + un[0:-2, 1:-1])) + 
-                        F * dt)
-
-        v[1:-1, 1:-1] = (vn[1:-1, 1:-1] -
-                        un[1:-1, 1:-1] * dt / dx * 
-                        (vn[1:-1, 1:-1] - vn[1:-1, 0:-2]) -
-                        vn[1:-1, 1:-1] * dt / dy * 
-                        (vn[1:-1, 1:-1] - vn[0:-2, 1:-1]) -
-                        dt / (2 * rho * dy) * 
-                        (p[2:, 1:-1] - p[0:-2, 1:-1]) +
-                        nu * (dt / dx**2 *
-                        (vn[1:-1, 2:] - 2 * vn[1:-1, 1:-1] + vn[1:-1, 0:-2]) +
-                        dt / dy**2 * 
-                        (vn[2:, 1:-1] - 2 * vn[1:-1, 1:-1] + vn[0:-2, 1:-1])))
-
-        # Periodic BC u @ x = 2     
-        u[1:-1, -1] = (un[1:-1, -1] - un[1:-1, -1] * dt / dx * 
-                    (un[1:-1, -1] - un[1:-1, -2]) -
-                    vn[1:-1, -1] * dt / dy * 
-                    (un[1:-1, -1] - un[0:-2, -1]) -
-                    dt / (2 * rho * dx) *
-                    (p[1:-1, 0] - p[1:-1, -2]) + 
-                    nu * (dt / dx**2 * 
-                    (un[1:-1, 0] - 2 * un[1:-1,-1] + un[1:-1, -2]) +
-                    dt / dy**2 * 
-                    (un[2:, -1] - 2 * un[1:-1, -1] + un[0:-2, -1])) + F * dt)
-
-        # Periodic BC u @ x = 0
-        u[1:-1, 0] = (un[1:-1, 0] - un[1:-1, 0] * dt / dx *
-                    (un[1:-1, 0] - un[1:-1, -1]) -
-                    vn[1:-1, 0] * dt / dy * 
-                    (un[1:-1, 0] - un[0:-2, 0]) - 
-                    dt / (2 * rho * dx) * 
-                    (p[1:-1, 1] - p[1:-1, -1]) + 
-                    nu * (dt / dx**2 * 
-                    (un[1:-1, 1] - 2 * un[1:-1, 0] + un[1:-1, -1]) +
-                    dt / dy**2 *
-                    (un[2:, 0] - 2 * un[1:-1, 0] + un[0:-2, 0])) + F * dt)
-
-        # Periodic BC v @ x = 2
-        v[1:-1, -1] = (vn[1:-1, -1] - un[1:-1, -1] * dt / dx *
-                    (vn[1:-1, -1] - vn[1:-1, -2]) - 
-                    vn[1:-1, -1] * dt / dy *
-                    (vn[1:-1, -1] - vn[0:-2, -1]) -
-                    dt / (2 * rho * dy) * 
-                    (p[2:, -1] - p[0:-2, -1]) +
-                    nu * (dt / dx**2 *
-                    (vn[1:-1, 0] - 2 * vn[1:-1, -1] + vn[1:-1, -2]) +
-                    dt / dy**2 *
-                    (vn[2:, -1] - 2 * vn[1:-1, -1] + vn[0:-2, -1])))
-
-        # Periodic BC v @ x = 0
-        v[1:-1, 0] = (vn[1:-1, 0] - un[1:-1, 0] * dt / dx *
-                    (vn[1:-1, 0] - vn[1:-1, -1]) -
-                    vn[1:-1, 0] * dt / dy *
-                    (vn[1:-1, 0] - vn[0:-2, 0]) -
-                    dt / (2 * rho * dy) * 
-                    (p[2:, 0] - p[0:-2, 0]) +
-                    nu * (dt / dx**2 * 
-                    (vn[1:-1, 1] - 2 * vn[1:-1, 0] + vn[1:-1, -1]) +
-                    dt / dy**2 * 
-                    (vn[2:, 0] - 2 * vn[1:-1, 0] + vn[0:-2, 0])))
-
-
-        # Wall BC: u,v = 0 @ y = 0,2
-        u[0, :] = 0
-        u[-1, :] = 0
-        v[0, :] = 0
-        v[-1, :]=0
-        
-        u_list.append(u.copy())
-        v_list.append(v.copy())
-        p_list.append(p.copy())
-
-    u_list = np.stack(u_list)
-    v_list = np.stack(v_list)
-    p_list = np.stack(p_list)
-
-    return u_list, v_list, p_list
-
-
-if __name__ == "__main__":
-    from matplotlib import pyplot as plt, cm
-    from mpl_toolkits.mplot3d import Axes3D
-
-    import argparse
-    parser = argparse.ArgumentParser()
-    parser.add_argument('--nx', type=int, default=41,
-                        help='grid size along nx [default: 41]')
-    parser.add_argument('--ny', type=int, default=41,
-                        help='grid size along ny [default: 41]')
-    parser.add_argument('--dt', type=float, default=.001,
-                        help='stepsize over time [default: .001]')
-    args = parser.parse_args()
-
-    dirname = os.path.dirname(__file__)
-    os.makedirs(os.path.join(dirname, 'images_periodic'), exist_ok=True)
-    os.makedirs(os.path.join(dirname, 'data_periodic'), exist_ok=True)
-
-    nx = args.nx
-    ny = args.ny
-    nt = 200
-    nit = 50
-    dx = 2 / (nx - 1)
-    dy = 2 / (ny - 1)
-    x = np.linspace(0, 2, nx)
-    y = np.linspace(0, 2, ny)
-    X, Y = np.meshgrid(x, y)
-
-    rho = 1
-    nu = .1
-    F = 1  # a source term
-    dt = args.dt
-
-    u = np.zeros((ny, nx))
-    v = np.zeros((ny, nx))
-    p = np.zeros((ny, nx))
-    b = np.zeros((ny, nx))
-
-    u_dset, v_dset, p_dset = navier_stokes_flow(nt, u, v, dt, dx, dy, p, rho, nu, F)
-    u, v, p = u_dset[-1], v_dset[-1], p_dset[-1]
-
-    fig = plt.figure(figsize=(11, 7), dpi=100)
-    plt.contourf(X, Y, p, alpha=0.5, cmap=cm.viridis)
-    plt.colorbar()
-    plt.contour(X, Y, p, cmap=cm.viridis)
-    plt.quiver(X[::2, ::2], Y[::2, ::2], u[::2, ::2], v[::2, ::2])
-    plt.xlabel('X')
-    plt.ylabel('Y')
-    plt.savefig('images_periodic/quiver_nx_{}_ny_{}_dt_{}.pdf'.format(nx, ny, dt))
-
-    fig = plt.figure(figsize=(11, 7), dpi=100)
-    plt.contourf(X, Y, p, alpha=0.5, cmap=cm.viridis)
-    plt.colorbar()
-    plt.contour(X, Y, p, cmap=cm.viridis)
-    plt.streamplot(X, Y, u, v)
-    plt.xlabel('X')
-    plt.ylabel('Y')
-    plt.savefig('images_periodic/streamplot_nx_{}_ny_{}_dt_{}.pdf'.format(nx, ny, dt))
-
-    np.savez('data_periodic/data_nx_{}_ny_{}_dt_{}.npz'.format(nx, ny, dt),
-             X=X, Y=Y, p=p_dset, u=u_dset, v=v_dset)
